21. BBS 파일 삭제기능 구현
 - ~p612

+++ src/main/java/org/zerock/controller/UploadController.java
@@ -21,6 +21,7 @@ import org.springframework.ui.Model;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.zerock.util.MediaUtils;
@@ -35,24 +36,30 @@ public class UploadController {
 	
 	private static final Logger logger =LoggerFactory.getLogger(UploadController.class);
 	
+	
 	@ResponseBody
-	@RequestMapping(value="/deleteFile",method=RequestMethod.POST)
-	public ResponseEntity<String> deleteFile(String fName)throws Exception{
+	@RequestMapping(value="/deleteAllFiles",method=RequestMethod.POST)
+	public ResponseEntity<String> deleteFile(@RequestParam("files[]")String[] files)throws Exception{
 		
-		String fileName = URLDecoder.decode(fName, "UTF-8") ; //한글 인코딩
-		logger.info("delete file: "+ fName);
-		logger.info("After encoding delete file: "+ fileName);
-		String formatName = fileName.substring(fileName.lastIndexOf(".")+1);
-		MediaType mType = MediaUtils.getMediaType(formatName);
+		logger.info("delete all files: "+files);
 		
-		if(mType != null){
-			
-			String front = fileName.substring(0,12);
-			String end = fileName.substring(14);
-			new File(uploadPath +(front + end).replace('/', File.separatorChar)).delete();
-			
+		if(files ==null  || files.length ==0 ) {
+			return new ResponseEntity<String>("deleted",HttpStatus.OK); 
 		}
-		new File(uploadPath + fileName.replace('/', File.separatorChar)).delete();
+		
+		for(String fileName : files) {
+			String formatName = fileName.substring(fileName.lastIndexOf(".")+1);
+			MediaType mType = MediaUtils.getMediaType(formatName);
+			
+			if(mType != null){
+				
+				String front = fileName.substring(0,12);
+				String end = fileName.substring(14);
+				new File(uploadPath +(front + end).replace('/', File.separatorChar)).delete();
+				
+			}
+			new File(uploadPath + fileName.replace('/', File.separatorChar)).delete();
+		}
 		
 		return new ResponseEntity<String>("deleted",HttpStatus.OK);
 	}
